Compile whole legacy module
